Provide Fahrenheit representation of temperature value
Note:
From the caller's perspective, this is not a conversion (even though
in the unit test it is used like a conversion to be able to test it).
We want to be able to treat the temperature object as just an object,
pass it around and use it, and request one of the supported represen-
tations on demand (so far Celsius and Fahrenheit).

Why not `toFahrenheit`?
=> We do not want the object to act like a "converter" for
temperature measurements. The object itself represents a certain
temperature, which can then be represented on different supported
scales. A temperature in and of itself does not exist on a "scale",
let alone any specific one.

We are implementing the "conversion" (note that the conversion is
completely encapsulated) ourselves instead of looking up a formula.
Remapping values is a common programming task, and sometimes a simple
"formula" may not be available. Physical formulas for temperature
conversions are just specific instances of the remap function with
predefined coefficients. We can derive the "formula" (i.e. the
coefficients for remapping the scales) from a few known values on
each scale, as long as equivalent values (i.e. the same temperature
on both scales) are known.