Add additional test tuples to prevent false positives
We now test against all 3 tuples of known values. Otherwise, with
some (rather low) likelihood, we might have a false positive when
testing only 2 values where potentially an offset would be enough
without a scale factor to convert from one scale to the other (as
is possible between Kelvin and Celsius, but not between Celsius and
Kelvin). E.g. to convert from 0 °C to 32 °F, it would theoretically
be sufficient to add 32 °F, but not with any other pair of equi-
valent values on these scales.