Simplify tests (note: read below)
This adds some redundancy: We can now remove the 2 tests
"shouldPreserveValue_fromCelsiusToCelsius" and
"shouldPreserveValue_fromFahrenheitToFahrenheit", because this
change covers all combinations. Normally we could remove them already
with this commit, but in this case we will do it in the next commit
for demonstration purposes.

How do we know that refactoring our tests like this is safe? Who
tests the tests?
=> The old tests (before refactoring) test the new ones (after
refactoring), by being executed together. If everything still passes,
we can remove the redundancy and keep only the new, simplified/
refactored tests.