Use Kelvin instead of Celsius as internal representation
This change is not required at all, but Kelvin seems a good choice
as an internal representation since it is the most "technical" scale
and because it has the simplest lower bound. This also makes range
checks easier, e.g. in the constructor (which we will add soon),
since values below absolute zero should not result in a valid
temperature object (at least not by default; in special cases where
this should actually be allowed, it must be a well-justified
decision).

This change is a good example for how we can easily and confidently
change the implementation details if the object has a small surface
area and the tests do not depend on internals.